requirements.txt
# Core Deep Learning
torch>=2.0.0
torchvision>=0.15.0
torchaudio>=2.0.0

# Mamba State Space Models
mamba-ssm>=1.2.0
causal-conv1d>=1.2.0

# Medical Image Processing
SimpleITK>=2.3.0
nibabel>=5.0.0
medpy>=0.4.0

# Image Processing & Augmentation
albumentations>=1.3.0
opencv-python>=4.8.0
Pillow>=10.0.0
scikit-image>=0.21.0

# Scientific Computing
numpy>=1.24.0
scipy>=1.11.0
pandas>=2.0.0
scikit-learn>=1.3.0

# Visualization
matplotlib>=3.7.0
seaborn>=0.12.0
tensorboard>=2.14.0

# Configuration & Logging
PyYAML>=6.0
omegaconf>=2.3.0
wandb>=0.15.0
tqdm>=4.65.0

# Model Analysis
fvcore>=0.1.5
ptflops>=0.7.0
thop>=0.1.1

# Einops for tensor operations
einops>=0.6.0

# Timm for pretrained models
timm>=0.9.0

# Testing
pytest>=7.4.0

# Jupyter support
jupyter>=1.0.0
ipywidgets>=8.0.0
